make PointCloud a Geometry instead of an Object
