+++ Cargo.toml
@@ -37,6 +37,7 @@ wavefront_obj = { version = "10.0", optional = true }
 image = { version = "0.23", optional = true, default-features = false, features = ["gif", "jpeg", "ico", "png", "pnm", "tga", "tiff", "webp", "bmp", "hdr", "dxt", "dds", "farbfeld"]}
 egui = { version = "0.13", optional = true }
 ply-rs = { version = "0.1.3", optional = true }
+rand = "0.8.5"
 
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
