+++ examples/point_cloud/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "point_cloud"
+name = "point-cloud"
 version = "0.1.0"
 authors = ["Asger Nyman Christiansen <asgernyman@gmail.com>"]
 edition = "2021"
